Rename files from tpub to rpub
