+++ data/README.md
@@ -1,10 +1,10 @@
-kpub: data
+rpub: data
 ==========
 
 This directory contains the master copy of the SQLite database file,
-`kpub.db`.
+`rpub.db`.
 
 These are the sub-directories:
-- `output/` contains the markdown output that is created by running `kpub --save --month`.
-- `legacy/` contains the pre-2015 publication list and a script to extract the bibcodes  for ingestion into `kpub`.
+- `output/` contains the markdown output that is created by running `rpub --save --month`.
+- `legacy/` contains the pre-2015 publication list and a script to extract the bibcodes  for ingestion into `rpub`.
