+++ rpub/__init__.py
@@ -6,4 +6,4 @@ import os
 
 PACKAGEDIR = os.path.abspath(os.path.dirname(__file__))
 
-from .tpub import *
+from .rpub import *
