+++ setup.py
@@ -12,27 +12,27 @@ except (IOError, ImportError):
 
 # Command-line tools
 entry_points = {'console_scripts': [
-    'tpub = tpub:tpub',
-    'tpub-update = tpub:tpub_update',
-    'tpub-add = tpub:tpub_add',
-    'tpub-delete = tpub:tpub_delete',
-    'tpub-import = tpub:tpub_import',
-    'tpub-export = tpub:tpub_export',
+    'rpub = rpub:rpub',
+    'rpub-update = rpub:rpub_update',
+    'rpub-add = rpub:rpub_add',
+    'rpub-delete = rpub:rpub_delete',
+    'rpub-import = rpub:rpub_import',
+    'rpub-export = rpub:rpub_export',
     'tpub-plot = tpub:tpub_plot',
     'tpub-spreadsheet = tpub:tpub_spreadsheet'
 ]}
 
-setup(name='tpub',
+setup(name='rpub',
       version='1.1.0',
       description="A simple tool to keep track of the publications related "
                   "to NASA's TESS mission.",
       long_description=long_description,
-      author='Tom Barclay',
-      author_email='tom@tombarclay.com',
+      author='Rob Zellem',
+      author_email='robert.t.zellem@nasa.gov',
       license='MIT',
       url='',
-      packages=['tpub'],
-      data_files=[('tpub/templates', ['tpub/templates/template.md', 'tpub/templates/template-overview.md'])],
+      packages=['rpub'],
+      data_files=[('rpub/templates', ['rpub/templates/template.md', 'rpub/templates/template-overview.md'])],
       install_requires=["jinja2",
                         "six",
                         "astropy",
